requirements.txt
pur
pygithub
pyyaml
